+++ entropiaevents/__init__.py
@@ -1,18 +1,60 @@
-import requests
 import json
 import re
-
+import locale
+from datetime import datetime, timedelta
 from obelixtools import API
 
 class Event:
-    def __init__(self, date, time, location, description):
-        self.date = date
-        self.time = time
+    def __init__(self, date_str, time_str, location, description):
+        self._defaults = dict(
+            locale='de_DE',
+            duration=timedelta(hours=2)
+        )
+        self._parse_date(date_str, time_str)
         self.location = location
         self.description = description
 
+    @property
+    def start_date(self):
+        if not self._dates:
+            return 'Date parsing error!'
+        return self._dates[0].strftime('%x %X')
+
+    @property
+    def end_date(self):
+        if not self._dates[1]:
+            return (self._dates[0] + self.default_event_duration).strftime('%x %X')
+        return self._dates[1].strftime('%x %X')
+
+    def set_locale(self, value='de_DE'):
+        self._defaults['locale'] = value;
+
+    @property
+    def locale(self):
+        return self._defaults['locale']
+
+    def set_default_event_duration(self, hours=2):
+        self._defaults['duration'] = timedelta(hours=hours)
+
+    @property
+    def default_event_duration(self):
+        return self._defaults['duration']
+
+    def _parse_date(self, date_str, time_str):
+        locale.setlocale(locale.LC_TIME, self.locale)
+        date_str = date_str.split('-')
+        if len(date_str) == 2:
+            fmt = '%a, %d.%m.%Y'
+            try:
+                self._dates = [datetime.strptime(el.strip(), fmt) for el in date_str]
+            except:
+                self._dates = False
+        else:
+            full_date = date_str[0].strip() + ' ' + time_str.strip()
+            self._dates = [datetime.strptime(full_date, '%a, %d.%m.%Y ab %H:%M'), False]
+
     def __str__(self):
-        return u'{date} {time} - {location} - {description}'.format(**self.__dict__)
+        return u'{} : {} - {location} - {description}'.format(self.start_date, self.end_date, **self.__dict__)
 
 class WikiEvents:
     API_URL = 'https://entropia.de/api.php?format=json&action=parse&page=Vorlage:Termine'